Add files via upload
n this block of code, we define two functions. The create_ticket function creates a new ticket and stores it in the database. The send_email function sends a confirmation email to the user with the ticket ID. We use the smtplib package to send the email over the SMTP protocol.